+++ plone/app/widgets/tests/test_at.py
@@ -343,7 +343,6 @@ class RelatedItemsWidgetTests(unittest.TestCase):
                 'value': '{};{}'.format(IUUID(obj1), IUUID(obj2)),
                 'pattern': 'relateditems',
                 'pattern_options': {
-                    'folderTypes': ['SomeType'],
                     'selectableTypes': ['SomeSelectableType', ],
                     'homeText': u'Home',
                     'searchAllText': u'Entire site',
@@ -388,7 +387,6 @@ class RelatedItemsWidgetTests(unittest.TestCase):
                 'value': '{}'.format(IUUID(obj1)),
                 'pattern': 'relateditems',
                 'pattern_options': {
-                    'folderTypes': ['SomeType'],
                     'homeText': u'Home',
                     'separator': ';',
                     'orderable': True,
@@ -421,7 +419,6 @@ class RelatedItemsWidgetTests(unittest.TestCase):
                 'value': '',
                 'pattern': 'relateditems',
                 'pattern_options': {
-                    'folderTypes': ['SomeType'],
                     'homeText': u'Home',
                     'separator': ';',
                     'orderable': True,
@@ -471,7 +468,6 @@ class RelatedItemsWidgetTests(unittest.TestCase):
                 'value': '{}'.format(IUUID(obj1)),
                 'pattern': 'relateditems',
                 'pattern_options': {
-                    'folderTypes': ['SomeType'],
                     'homeText': u'Home',
                     'separator': ';',
                     'orderable': True,
@@ -502,7 +498,6 @@ class RelatedItemsWidgetTests(unittest.TestCase):
                 'value': '{};{}'.format(IUUID(obj1), IUUID(obj2)),
                 'pattern': 'relateditems',
                 'pattern_options': {
-                    'folderTypes': ['SomeType'],
                     'homeText': u'Home',
                     'separator': ';',
                     'orderable': True,